README.md update
It will be easier to force set lootquality of the weapons (400,500,600) rewards of the lasts classes challenges using the same system as the class selection suitcases.
So basically the 3 lasts challenges will grant you a key and a weapSelection item. When you choose an AK47 suitcase, it will countain the weapon and some ammo boxes.